Added documentation for `tuning_method` and `seed` attributes
`tuning_method`: This brings an effective Hyperparameter Tuning strategy to a declarative level where the user can specify (should they want to tune hyper parameters for their downstream model) if they wish to use 'TrainValidationSplit' or 'CrossValidator' as their tuning strategy. Combining this with `extra_config`, the user is expected to have nearly total control in the parameter tuning process, by use their preferred tuning strategy with the grid of their choice. ('TrainValidationSplit' may be more recommended for large datasets though.)

`seed`: This brings the random seeding of user's workload to a declarative level as well, where the user may specify a particular seed to be used in the entire model selection (and hyperparameter tuning, if applicable) process. This will enable result reproducibility, an important component in Machine Learning.